+++ collective/searchandreplace/__init__.py
@@ -33,7 +33,10 @@ __version__ = '$ Revision 0.0 $'[11:-2]
 
 from zope.i18nmessageid import MessageFactory
 
-
+# Define permission string, so we can import it if needed.  See also
+# the permission in configure.zcml.
+UseSearchAndReplace = "collective.searchandreplace: Use Search And Replace"
+# Message factory:
 SearchAndReplaceMessageFactory = MessageFactory('SearchAndReplace')